Set VIRTUALENVWRAPPER_PYTHON=$(pyenv-which python) (if unset)
This appears to be necessary, because when loading
virtualenvwrapper/pyenv-virtualenvwrapper after pyenv, it would use
~/.pyenv/shims/python and make it necessary to install virtualenvwrapper
in every virtualenv to make the hooks work.

The reasons for this are basically the same as for
VIRTUALENVWRAPPER_SCRIPT, I suppose.
Although this gets still set unconditionally always (e.g. in a tmux
subshell).